fix: ts error
